+++ connecteur-type/SEDAConnecteur.class.php
@@ -56,10 +56,4 @@ abstract class SEDAConnecteur extends Connecteur
             );
         }
     }
-
-    public function getTransferId(string $bordereau): string
-    {
-        $xml = \simplexml_load_string($bordereau);
-        return (string)($xml->TransferIdentifier ?? $xml->MessageIdentifier);
-    }
 }